+++ app/views/admin/jobs/edit.html.erb
@@ -1,0 +1,29 @@
+<h2>更新招聘信息</h2>
+<div class="col-md-4 col-md-offset-4">
+	<hr>
+	<%= form_for [:admin, @job] do |f| %>
+	标题
+	<%= f.text_field :title %>
+	<br>
+
+	描述内容
+	<%= f.text_area :description %>
+	<br>
+
+	最低工资
+	<%= f.text_field :min_salary %>
+	<br>
+
+	最高工资
+	<%= f.text_field :max_salary %>
+	<br>	
+
+	联系信息
+	<%= f.text_field :contact_info %>
+	<br>
+
+	<%= f.submit "更新信息", :disable_with => "提交中..." %>
+	<% end %>
+
+
+</div>
