+++ app/views/common/_navbar.html.erb
@@ -18,6 +18,10 @@
                           <b class="caret"></b>
                         </a>  
                         <ul class="dropdown-menu">
+                            <% if current_user.admin? %>
+                                <li><%= link_to("管理员后台", admin_jobs_path) %></li>
+                                <li class="divider"> </li>
+                            <% end %>
                             <li> <%= link_to("登出", destroy_user_session_path, method: :delete) %> </li>
                         </ul>
                     </li>
