+++ app/views/jobs/index.html.erb
@@ -1,5 +1,4 @@
 <div class="col-md-12">
-  <%= link_to("发布新的招聘信息", new_job_path, class:"btn btn-primary pull-right") %>  
   
   <table class="table table-hover">
     <thead>
@@ -8,9 +7,6 @@
         <td>工作</td>
         <td>薪酬</td>
         <td>内容</td>
-        <% if current_user %>
-        <td>操作</td>
-        <% end %>
       </tr>
     </thead>
     
@@ -21,12 +17,6 @@
           <td><%= job.title %></td>
           <td><%= "#{job.min_salary} ~ #{job.max_salary}"%></td>
           <td><%= job.description %></td>
-          <% if current_user %>
-            <td>
-              <%= link_to("编辑", edit_job_path(job), class: "btn btn-sm btn-default") %>
-              <%= link_to("删除", job_path(job), class: "btn btn-sm btn-default", method: :delete, data: { confirm: "确认删除信息？"}) %>
-            </td>
-          <% end %>
         </tr>
       <% end %>
     </tbody>
