+++ config/routes.rb
@@ -3,5 +3,9 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   resources :jobs
 
+  namespace :admin do
+  	resources :jobs
+  end
+
   root "jobs#index"
 end
