+++ test/controllers/admin/jobs_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::JobsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
